Add Role Table/CRUD

+++ app/routers/user.py
@@ -13,6 +13,7 @@ from app.helpers.auth import (
     check_password,
     create_access_token,
     get_current_user,
+    get_current_user_role,
 )
 from fastapi.security import OAuth2PasswordBearer, OAuth2PasswordRequestForm
 from app.models.user import UserManager
@@ -48,7 +49,10 @@ def read_user(
     user_id: int,
     current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
+    current_user_role: str = Depends(get_current_user_role),
 ):
+    if current_user.id != user_id and current_user_role != "admin":
+        raise HTTPException(status_code=403, detail="Permission denied")
     db_user = get_user(db, user_id=user_id)
     if db_user is None:
         raise HTTPException(status_code=404, detail="User not found")
@@ -61,7 +65,10 @@ def read_users(
     limit: int = 10,
     current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
+    current_user_role: str = Depends(get_current_user_role),
 ):
+    if current_user_role != "admin":
+        raise HTTPException(status_code=403, detail="Permission denied")
     users = get_users(db, skip=skip, limit=limit)
     return users
 
@@ -70,9 +77,12 @@ def read_users(
 def update_user_route(
     user_id: int,
     user_update: UserUpdate,
-    current_user: User = Depends(lambda db: get_current_user(db)),
+    current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
+    current_user_role: str = Depends(get_current_user_role),
 ):
+    if current_user.id != user_id and current_user_role != "admin":
+        raise HTTPException(status_code=403, detail="Permission denied")
     db_user = update_user(db, user_id=user_id, user=user_update)
     if db_user is None:
         raise HTTPException(status_code=404, detail="User not found")
@@ -84,7 +94,10 @@ def delete_user_route(
     user_id: int,
     current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db),
+    current_user_role: str = Depends(get_current_user_role),
 ):
+    if current_user_role != "admin":
+        raise HTTPException(status_code=403, detail="Permission denied")
     success = delete_user(db, user_id=user_id)
     if not success:
         raise HTTPException(status_code=404, detail="User not found")
